requirements.txt
asgiref==3.11.1
boto3==1.42.62
django==5.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-storages==1.14.6
gunicorn>=21.2.0
drf-spectacular==0.28.0
pillow==12.1.1
psycopg2-binary==2.9.11
pyjwt==2.11.0
python-decouple==3.8
sqlparse==0.5.5
